[fix] Change python version checks

+++ docker/build-wheels.sh
@@ -17,15 +17,14 @@ fi
 cd /opt/aimrocks
 
 echo "build python wheels"
-python_versions=("cp36-cp36m" "cp37-cp37m" "cp38-cp38" "cp39-cp39")
-if [[ $platform != centos_5 ]]
-then
-  python_versions+=("cp310-cp310")
-  python_versions+=("cp311-cp311")
-fi
+python_versions=("cp36-cp36m" "cp37-cp37m" "cp38-cp38" "cp39-cp39" "cp310-cp310" "cp311-cp311")
 
 for python_version in "${python_versions[@]}"
 do
-  /opt/python/${python_version}/bin/python -m build
-  rm -rf build
+  python_exe=/opt/python/${python_version}/bin/python
+  if [ -f "$python_exe" ]
+  then
+    $python_exe -m build
+    rm -rf build
+  fi
 done
